adding new openGL rendere for dino run
