+++ app/src/main/cpp/dino_run/dino_run.cpp
@@ -1,0 +1,60 @@
+//
+// Created by Tristan on 2025-03-11.
+//
+
+#include <GLES2/gl2.h>
+#include <GLES2/gl2ext.h>
+#include <android/log.h>
+#include <jni.h>
+#include <math.h>
+#include <stdio.h>
+#include <stdlib.h>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+extern "C"
+JNIEXPORT void JNICALL
+Java_com_example_clicker_presentation_minigames_dinoRun_DinoRunJNI_init(JNIEnv *env, jobject thiz,
+                                                                        jint width, jint height) {
+
+}
